Labels einer Variablen wiedergeben.

Funktion zeigt SPSS-Labels einer Variablen an.

GetLabels(x, var, curr = TRUE)

Arguments

x
MZ Daten - Output von Funktion ImportData oder IndivImportData.

var
Character Vektor mit Variable(n) fuer die die Labels angezeigt werden sollen.

curr
TRUE/FALSE ob var auf "current" oder "previous" Zeitpunkt bezogen sein soll, d.h. falls comp_diff_lag ungleich NULL in der Funktion ImportData, koennen mit curr=FALSE auch die Labels des weniger aktuellen der beiden Datensaetze ausgegeben werden. curr=TRUE bezieht sich auf den aktuelleren der beiden Zeitpunkte.

Value

Output sind die Labels einer Variablen var falls im dg7-Datensatz welche definiert wurden.

See also

Examples

# Daten laden (oder Daten einlesen mit ImportData() bzw. IndivImportData())
data(mzTestData)

# Labels von Variable xerwstat abfragen
GetLabels(mzTestData,"xerwstat")

#> $xerwstat
#> $xerwstat$label
#> [1] "Activity status during the reference week"
#> 
#> $xerwstat$value.labels
#> not applicable (child less than 15 years) 
#>                                         9 
#>  compulsory military or community service 
#>                                         4 
#>                     economically inactive 
#>                                         3 
#>                                unemployed 
#>                                         2 
#>                             in employment 
#>                                         1 
#> 
#> 

# Labels von Variablen dtstd und xnuts2 abfragen
GetLabels(mzTestData,c("dtstd","xnuts2"))

#> $dtstd
#> $dtstd$label
#> [1] "Number of hours actually worked during the reference week in the main job"
#> 
#> $dtstd$value.labels
#> Filter 
#>     -3 
#> 
#> 
#> $xnuts2
#> $xnuts2$label
#> [1] "NUTS-2 regions (Austrian federal states)"
#> 
#> $xnuts2$value.labels
#>    Vorarlberg         Tyrol      Salzburg Upper Austria        Styria 
#>            34            33            32            31            22 
#>     Carinthia        Vienna Lower Austria    Burgenland 
#>            21            13            12            11 
#> 
#>