Schaetz- und Fehlerrechnungsfunktionen.

Funktion berechnet absolute (GroupSize) oder relative (GroupRate) Schaetzwerte und die zugehoerigen Fehler.

GroupRate(x, TFstring, TFstring2 = NULL, each = NULL, byeach = TRUE,
  thousands_separator = TRUE, digits = 2, replicates = FALSE)

GroupSize(x, TFstring = NULL, each = NULL,
  thousands_separator = TRUE, digits = 2, replicates = FALSE)

Arguments

x
MZ Daten - Output von Funktion ImportData.

TFstring
Character oder NULL: Logische Einschraenkung der Gruppe, im Fall der Anteile (GroupRate): Einschraenkung fuer Zaehler. Falls NULL, gilt keine Einschraenkung.

TFstring2
Character oder NULL: Logische Einschraenkung fuer den Nenner. Falls NULL, gilt keine Einschraenkung fuer den Nenner ausser wenn each ungleich NULL und byeach=TRUE.

each
Character oder NULL: Name der Variable nach der getrennt berechnet werden soll. Hier koennen auch mehrere Variablen angegeben werden und zwar in einem character string, getrennt durch ein +, siehe Examples.

byeach
Logical: Vorgabe fuer den Nenner. byeach bezieht sich auf den Parameter each und gibt an, ob bei GroupRate zusaetzlich zu TFstring2 auch die jeweilige durch each definierte Einschraenkung in den Nenner kommen soll, also ob die Raten jeweils bezogen auf die Auspraegungen bzw. Auspraegungskombinationen von each berechnet werden sollen (byeach=TRUE) oder rein bezogen auf TFstring2 (byeach=FALSE).

thousands_separator
Logical: Wenn TRUE, werden Tausendertrennzeichen angezeigt.

digits
Numerischer Wert: Anzahl der Nachkommastellen im angezeigten Ergebnis. Default ist 2.

replicates
Fürge einen Vektor aus Schätzwerten zum Output hinzu? Die Anzahl der Schätzwerte pro Gruppe in each entspricht der Anzahl der Bootstrapreplikate (typischerweise 500). Siehe auch getReplicates.

Value

Output ist ein Objekt der Klasse mzR.

Details

Wiedergegeben wird der Schaetzer est, der Stichprobenfehler sd, der Variationskoeffizient cv und die untere/obere Grenze des 95% Konfidenzintervalls cil_2.5%/ciu_97.5%. Die Fehler werden mit Hilfe von Replikationsgewichten aus einem Bootstrapverfahren berechnet, d.h. sd entspricht der Standardabweichung der mit den Bootstrapgewichten berechneten Schaetzwerte, cil_2.5% und ciu_97.5% sind die entsprechenden 2.5% und 97.5% Quantile und cv=sd/est.

See also

Examples

# Daten laden (oder Daten einlesen mit ImportData() bzw. IndivImportData())
data(mzTestData)

# Arbeitslosenzahlen: Absolutwerte und Veraenderung
GroupSize(mzTestData,TFstring="xerwstat==2&balt>=15&balt<=74")

#>                        est          sd          cv    cil_2.5%    ciu_97.5%
#> year2           258,847.97   185,932.8        0.72   83,024.03    520,470.4
#> year1           336,981.66   355,645.5        1.06    5,364.08    801,720.5
#> Absolute change -78,133.69   340,417.5       -4.36 -356,939.05    387,517.8
#> Relative change     -23.19 9,513,740.1 -410,316.71      -81.31 19,146,156.7

# Arbeitslosenquoten: Prozentwerte und Veraenderung 
GroupRate(mzTestData,TFstring="xerwstat==2&balt>=15&balt<=74",
          TFstring2="xerwstat%in%c(1,2)&balt>=15&balt<=74")

#>                    est           sd          cv cil_2.5%     ciu_97.5%
#> year2             5.89         3.82        0.65     1.73         10.44
#> year1             6.58         7.73        1.17     0.09         17.79
#> Absolute change  -0.69         7.33      -10.61    -7.96          8.14
#> Relative change -10.50 8,966,973.30 -854,080.27   -80.91 18,045,846.98

# Oesterreichische Bevoelkerung nach Bundesland und Geschlecht
GroupSize(mzTestData,TFstring=NULL,each="xnuts2+bsex")

#> $xnuts2_bsex_11001
#>                        est        sd    cv    cil_2.5%  ciu_97.5%
#> year2           185,263.10 21,948.62  0.12  156,969.78 205,504.42
#> year1           194,532.10 42,980.83  0.22  193,080.74 283,450.54
#> Absolute change  -9,269.00 53,719.94 -5.80 -124,227.10  -6,745.15
#> Relative change      -4.76     18.28 -3.84      -43.83      -3.29
#> 
#> $xnuts2_bsex_11002
#>                       est           sd         cv  cil_2.5%    ciu_97.5%
#> year2           98,937.14    21,946.62       0.22 78,695.82   127,230.46
#> year1           88,899.48    42,980.83       0.48      1.83    90,371.63
#> Absolute change 10,037.66    53,716.79       5.35  7,507.52   124,974.97
#> Relative change     11.29 3,396,203.88 300,787.98      9.63 7,285,517.30
#> 
#> $xnuts2_bsex_12001
#>                        est         sd    cv   cil_2.5%  ciu_97.5%
#> year2           561,527.58 308,543.09  0.55        2.6 692,055.89
#> year1           616,325.37  75,219.29  0.12  509,444.8 697,092.16
#> Absolute change -54,797.79 314,910.71 -5.75 -645,931.3   9,779.67
#> Relative change      -8.89      48.76 -5.48     -100.0       1.17
#> 
#> $xnuts2_bsex_12002
#>                          est         sd   cv   cil_2.5%    ciu_97.5%
#> year2           1,049,565.51 308,555.08 0.29 919,028.57 1,611,090.49
#> year1             985,793.66  75,219.29 0.08 904,964.62 1,092,611.99
#> Absolute change    63,771.84 314,930.51 4.94    -751.98   654,967.65
#> Relative change         6.47      32.91 5.09      -0.18        68.51
#> 
#> $xnuts2_bsex_13001
#>                          est         sd    cv      cil_2.5%    ciu_97.5%
#> year2             608,973.37 248,745.42  0.41    414,042.83   938,551.08
#> year1           1,179,773.90 276,865.07  0.23  1,031,807.35 1,681,767.49
#> Absolute change  -570,800.53 468,956.99 -0.82 -1,173,603.78   -94,634.18
#> Relative change       -48.38      29.11 -0.60        -70.38        -9.12
#> 
#> $xnuts2_bsex_13002
#>                          est           sd         cv   cil_2.5%    ciu_97.5%
#> year2           1,138,247.65    248,728.8       0.22 808,669.94  1,333,168.8
#> year1             539,146.41    276,865.1       0.51  37,278.92    687,239.1
#> Absolute change   599,101.23    468,925.8       0.78 122,808.79  1,201,694.2
#> Relative change       111.12 25,258,113.1 227,304.17      18.04 50,831,081.9
#> 
#> $xnuts2_bsex_21001
#>                        est         sd   cv    cil_2.5%  ciu_97.5%
#> year2           249,421.08 108,036.75 0.43  165,650.40 415,049.76
#> year1           210,938.63 100,366.94 0.48  103,573.58 321,464.30
#> Absolute change  38,482.45 187,798.36 4.88 -110,990.57 311,476.18
#> Relative change      18.24     150.82 8.27      -39.94     302.99
#> 
#> $xnuts2_bsex_21002
#>                        est         sd    cv    cil_2.5%  ciu_97.5%
#> year2           299,699.54 108,026.88  0.36  134,070.86 383,443.77
#> year1           337,389.91 100,366.94  0.30  226,842.93 444,733.65
#> Absolute change -37,690.37 187,789.47 -4.98 -310,662.79 111,777.51
#> Relative change     -11.17      49.64 -4.44      -69.82      41.42
#> 
#> $xnuts2_bsex_22001
#>                        est         sd   cv   cil_2.5%    ciu_97.5%
#> year2           781,989.89 171,938.56 0.22 643,442.26 1,024,927.07
#> year1           342,457.74 143,667.19 0.42 161,052.31   481,579.78
#> Absolute change 439,532.14 243,281.24 0.55 162,808.52   662,361.19
#> Relative change     128.35     177.27 1.38      33.81       433.09
#> 
#> $xnuts2_bsex_22002
#>                         est         sd    cv    cil_2.5%    ciu_97.5%
#> year2            417,610.63 171,898.09  0.41  174,779.68   556,158.26
#> year1            852,056.71 143,667.19  0.17  712,844.71 1,033,372.18
#> Absolute change -434,446.07 243,251.37 -0.56 -657,078.92  -157,632.49
#> Relative change      -50.99      25.73 -0.50      -78.38       -22.11
#> 
#> $xnuts2_bsex_31001
#>                        est        sd    cv    cil_2.5%  ciu_97.5%
#> year2           640,341.53  56,872.9  0.09  638,058.73 763,611.46
#> year1           678,107.56 233,969.1  0.35  445,567.52 942,262.50
#> Absolute change -37,766.03 269,355.3 -7.13 -301,220.37 285,689.14
#> Relative change      -5.57      42.6 -7.65      -31.89      64.51
#> 
#> $xnuts2_bsex_31002
#>                        est         sd   cv    cil_2.5%  ciu_97.5%
#> year2           769,421.99  56,872.86 0.07  646,152.06 771,705.76
#> year1           723,074.93 233,969.15 0.32  458,917.28 955,612.26
#> Absolute change  46,347.07 269,353.15 5.81 -277,096.63 309,804.11
#> Relative change       6.41      44.93 7.01      -28.97      67.87
#> 
#> $xnuts2_bsex_32001
#>                        est         sd   cv   cil_2.5%  ciu_97.5%
#> year2           392,769.95 121,374.41 0.31 190,964.21 461,358.69
#> year1           306,488.93  50,392.43 0.16 167,781.30 294,485.17
#> Absolute change  86,281.02 151,964.48 1.76 -96,861.44 253,384.79
#> Relative change      28.15      65.80 2.34     -30.95     121.95
#> 
#> $xnuts2_bsex_32002
#>                        est         sd    cv    cil_2.5%  ciu_97.5%
#> year2           134,841.37 121,374.94  0.90   66,252.63 336,647.47
#> year1           217,875.43  50,392.43  0.23  229,897.79 356,601.66
#> Absolute change -83,034.06 151,964.55 -1.83 -250,156.43 100,090.16
#> Relative change     -38.11      54.82 -1.44      -79.01      46.76
#> 
#> $xnuts2_bsex_33001
#>                        est        sd   cv  cil_2.5% ciu_97.5%
#> year2           531,313.22 50,834.64 0.10 505,046.2 620,545.6
#> year1           413,534.55 72,367.26 0.17 292,993.3 472,366.6
#> Absolute change 117,778.68 48,973.18 0.42 115,265.6 227,492.0
#> Relative change      28.48     21.70 0.76      24.6      77.3
#> 
#> $xnuts2_bsex_33002
#>                         est        sd    cv    cil_2.5%   ciu_97.5%
#> year2            182,404.49 50,844.90  0.28   93,172.07  208,707.95
#> year1            293,971.52 72,367.26  0.25  235,164.52  414,537.88
#> Absolute change -111,567.03 48,977.05 -0.44 -221,305.48 -109,079.08
#> Relative change      -37.95     11.35 -0.30      -67.05      -40.39
#> 
#> $xnuts2_bsex_34001
#>                        est        sd   cv   cil_2.5% ciu_97.5%
#> year2           169,206.60 37,197.94 0.22 199,813.18 280,924.3
#> year1           142,175.23 73,098.13 0.51  39,132.53 207,548.8
#> Absolute change  27,031.37 95,656.19 3.54  -3,764.18 214,309.6
#> Relative change      19.01    183.22 9.64       6.90     429.1
#> 
#> $xnuts2_bsex_34002
#>                        est        sd    cv    cil_2.5%  ciu_97.5%
#> year2           203,315.38 37,194.27  0.18   91,599.76 172,708.80
#> year1           227,404.95 73,098.13  0.32  162,017.04 330,433.29
#> Absolute change -24,089.57 95,652.48 -3.97 -211,351.29   6,720.34
#> Relative change     -10.59     34.51 -3.26      -69.79       9.16
#> 

# NOT RUN {
############################################################################################
#   Zusaetzliche Beispiele fuer DatennutzerInnen der Mikrozensus-Arbeitskraefteerhebung:   #
############################################################################################
# Quartal und zugehoeriges Vorjahrsquartal einlesen (Funktion fuer STAT-interne Nutzer)
dat <- ImportData(year=2014,quarter=4, comp_diff_lag=4)

# Oesterreichische Bevoelkerung nach Bundesland und Geschlecht
GroupSize(dat,TFstring=NULL,each="xnuts2+bsex")

# Arbeitslosenzahlen: Absolutwerte und Veraenderung
GroupSize(dat,TFstring="xerwstat==2&balt>=15&balt<=74")
# Arbeitslosenquoten: Prozentwerte und Veraenderung 
GroupRate(dat,TFstring="xerwstat==2&balt>=15&balt<=74",
  TFstring2="xerwstat%in%c(1,2)&balt>=15&balt<=74")
# Arbeitslosenzahl vom aktuelleren der beiden Quartale
GroupSize(dat[1],TFstring="xerwstat==2&balt>=15&balt<=74")
# Arbeitslosenquote vom weniger aktuellen Quartal
GroupRate(dat[2],TFstring="xerwstat==2&balt>=15&balt<=74",
  TFstring2="xerwstat%in%c(1,2)&balt>=15&balt<=74")

# Absolutwerte und Veraenderung fuer jede Auspraegung von xerwstat eingeschraenkt auf 15-74-Jaehrige
GroupSize(dat,TFstring="balt>=15&balt<=74",each="xerwstat")
# Prozentwerte (bezogen auf Gesamtbevoelkerung) und Veraenderung 
# fuer jede Auspraegung von xerwstat eingeschraenkt auf 15-74-Jaehrige
GroupRate(dat,TFstring="balt>=15&balt<=74",each="xerwstat")
# Arbeitslosenzahlen: Absolutwerte und Veraenderung fuer jedes Bundesland
GroupSize(dat,TFstring="xerwstat==2&balt>=15&balt<=74",each="xnuts2")

# Arbeitslosenquote: Prozentwerte und Veraenderung pro Bundesland
GroupRate(dat,TFstring="xerwstat==2&balt>=15&balt<=74",
  TFstring2="xerwstat%in%c(1,2)&balt>=15&balt<=74",each="xnuts2")
# Arbeitslosenquote: Prozentwerte und Veraenderung pro Geschlecht
GroupRate(dat,TFstring="xerwstat==2&balt>=15&balt<=74",
  TFstring2="xerwstat%in%c(1,2)&balt>=15&balt<=74",each="bsex")
# Arbeitslosenquote: Prozentwerte und Veraenderung pro Bundesland X Geschlecht
GroupSize(dat,TFstring="xerwstat==2&balt>=15&balt<=74",each="xnuts2+bsex")
GroupRate(dat,TFstring="xerwstat==2&balt>=15&balt<=74",
  TFstring2="xerwstat%in%c(1,2)&balt>=15&balt<=74",each="xnuts2+bsex")

# Haushalte: Quartal und Vorquartal einlesen.
dat <- ImportData(year=2014,quarter=4, comp_diff_lag=1, hh=TRUE)
# Absolutwerte: Anzahl der Hauptmietwohnungen ohne gueltiger Kostenangabe.
GroupSize(dat,TFstring="wrecht==3")
GroupSize(dat,TFstring="wrecht2%in%c(1:3)")
# }