Individuelle Datenfiles zu bereits eingelesenen Mikrozensus-Daten mergen.

Funktion liest vom Nutzer zur Verfuegung gestellte Datenfiles ein und fuehrt diese mit bereits eingelesenen dg7-Mikrozensus-Daten zusammen.

ImportAndMerge(x, curr_inFile, prev_inFile = NULL, mergeBy = "asbper",
  whichVar = NULL)

Arguments

x
Output-Objekt der Funktion ImportData()

curr_inFile
Pfad der Datei die eingelesen werden soll (bezogen auf den aktuelleren der beiden Zeitpunkte falls prev_inFile ungleich NULL). Eingelesen werden koennen Files vom Typ .sav, .csv und .csv.gz.

prev_inFile
Falls ungleich NULL, Pfad der Datei die eingelesen werden soll (bezogen auf den weniger aktuellen Zeitpunkt). Eingelesen werden koennen Files vom Typ .sav, .csv und .csv.gz.

mergeBy
Character Vektor mit Variable(n) nach denen gemerged werden soll (default="asbper").

whichVar
Falls ungleich NULL, Character Vektor mit Variable(n) aus dem zur Verfuegung gestellten Datenfile die im Output-File enthalten sein sollen. Die ?brigen Variablen werden weggelassen. Default ist NULL, dabei werden alle Variablen behalten.

Value

Output ist eine Liste mit einem oder zwei Elementen, je nach Laenge des Outputs der Funktion ImportData() .

See also

Examples

# NOT RUN {
## Evt. Memory Limit erhoehen (max bei 32-bit R ist 4095)
#memory.limit(size=4095)
#
### Quartalsdaten
#
## dg7-Mikrozensus-Files einlesen
dat <- ImportData(year=2014,quarter=4)
## Pfad fuer individuelle Daten festlegen
curr_inFile <- "...pfad.../MeineDaten_2014q4.sav"
## Individuelle Daten ueber asbhh und apkz zu bereits eingelesenen 
## dg7-Mikrozensus-Daten mergen
dat_indiv <- ImportAndMerge(dat,curr_inFile=curr_inFile,mergeBy=c("asbhh","apkz"))
## dg7-Mikrozensus-Files einlesen: Quartal und zugehoeriges Vorjahrsquartal
dat <- ImportData(year=2014,quarter=4, comp_diff_lag=4)
## Pfad fuer individuelle Daten festlegen: aktuelleres Quartal
curr_inFile <- "...pfad.../MeineDaten_2014q4.sav"
## Pfad fuer individuelle Daten festlegen: weniger aktuelles Quartal
prev_inFile <- "...pfad.../MeineDaten_2013q4.sav"
## Individuelle Daten ueber asbper (default) zu bereits eingelesenen 
## dg7-Mikrozensus-Daten mergen
dat_indiv <- ImportAndMerge(dat,curr_inFile=curr_inFile,prev_inFile=prev_inFile)

### Jahresdaten
#
## dg7-Mikrozensus-Jahresdaten einlesen
dat <- ImportData(year=2014)
curr_inFile <- "...pfad.../MeineDaten_2014j.sav"
## Individuelle Daten ueber asbhh und apkz zu bereits eingelesenen 
## dg7-Mikrozensus-Jahresdaten mergen
dat_indiv <- ImportAndMerge(dat, curr_inFile=curr_inFile, mergeBy=c("asbhh","apkz","amonat"))
### dg7-Mikrozensus-Jahresdaten einlesen: Jahr und Vorjahr
dat <- ImportData(year=2014, comp_diff_lag=1)
## Pfad fuer individuelle Daten festlegen: aktuelleres Jahr
curr_inFile <- "...pfad.../MeineDaten_2014j.sav"
## Pfad fuer individuelle Daten festlegen: weniger aktuelles Jahr
prev_inFile  <- "...pfad.../MeineDaten_2013j.sav"
## Individuelle Daten ueber asbper (default) und amonat zu bereits eingelesenen
## dg7-Mikrozensus-Jahresdaten mergen
dat_indiv <- ImportAndMerge(dat, curr_inFile=curr_inFile, mergeBy=c("asbper","amonat"))
# }