Mikrozensus-Files und zugehoerige Bootstrapgewichte einlesen (hausintern).
Source: R/ImportData.R

Funktion liest Mikrozensus-Files (dg7) und zugehoerige Bootstrapgewichte ein (basierend auf STAT-Filemanagement, d.h. diese Funktion funktioniert nur STAT-intern).

ImportData(year = NULL, quarter = NULL, comp_diff_lag = NULL,
  from = NULL, to = NULL, hh = FALSE, families = FALSE,
  whichVar = NULL, nbw = NULL, weightDecimals = 2)

Arguments

year
Numerischer Wert (Jahr).

quarter
Numerischer Wert (Quartal) oder NULL. Falls NULL, wird das ganze Jahr eingelesen.

comp_diff_lag
Numerischer Wert oder NULL. Falls NULL, wird keine Fehlerrechnung fuer Veraenderungen zwischen zwei Zeitpunkten durchgefuehrt und daher auch kein zusaetzliches File eingelesen. Falls solche Differenzen berechnet werden sollen, muss der Time-Lag angegeben werden. Einheiten sind Quartale falls quarter ungleich NULL, sonst Jahre.

from
Numerischer Vektor mit Jahr und Quartal oder NULL. Falls ungleich NULL, wird hier der Startzeitpunkt uebergeben falls mehr als ein Quartal eingelesen werden soll.

to
Numerischer Vektor mit Jahr und Quartal oder NULL. Falls ungleich NULL, wird hier der Endzeitpunkt uebergeben falls mehr als ein Quartal eingelesen werden soll.

hh
TRUE/FALSE ob auf Haushaltsreferenzpersonen (bstell=0) eingeschraenkt werden soll.

families
TRUE/FALSE ob die 'Stellung zur Familienreferenzperson' auf 'Referenzperson' (xfstell=1) eingeschraenkt werden soll.

whichVar
Character (vector) oder NULL. Falls ungleich NULL, Character Vektor mit Variable(n) aus dem dg7-Mikrozensus-File die im Output-File enthalten sein sollen. Die uebrigen Variablen werden weggelassen. Default ist NULL, dabei werden alle Variablen behalten.

nbw
Numerischer Wert oder NULL. Falls ungleich NULL, Anzahl an Bootstrap-Gewichten die eingelesen werden soll. Default ist NULL, dabei werden alle verfuegbaren Bootstrap-Gewichte eingelesen.

weightDecimals
Numerischer Wert oder NULL. Anzahl der Nachkommastellen der Stichprobengewichte, gerundet nach SPSS RND Logik (0.5 bwz. -0.5 wird dabei immer "weg von 0" gerundet). Falls NULL, werden die Gewichte nicht gerundet.

Value

Output ist eine Liste mit einem oder zwei Elementen, je nachdem ob comp_diff_lag=NULL oder nicht. Die Listenelemente sind Objekte der Klasse data.table.

See also

IndivImportData,GroupSize,GroupRate,Total,Mean,GetLabels,ImportAndMerge,export.mzR

Examples

# NOT RUN {
## Evt. Memory Limit erhoehen (max bei 32-bit R ist 4095)
#memory.limit(size=4095)
### Quartal
datq <- ImportData(year=2014, quarter=4)
### Jahr 
datj <- ImportData(year=2014)
### Quartal und zugehoeriges Vorjahrsquartal
datqdiff <- ImportData(year=2014, quarter=4, comp_diff_lag=4)
### Jahr und Vorjahr
datjdiff <- ImportData(year=2014, comp_diff_lag=1)
### Quartal und Vorquartal eingeschraenkt auf Haushaltsreferenzpersonen
datqdiffhh <- ImportData(year=2014, quarter=4, comp_diff_lag=1, hh=TRUE)
### Quartal eingeschraenkt auf Familienauswertungsrelevantes
datqfam <- ImportData(year=2014, quarter=4, families=TRUE)
### Mehr als ein Jahr einlesen (wegen Memory Limit derzeit begrenzt moeglich
### bzw. sollte 'whichVar' verwendet werden)
datzr <- ImportData(from=c(2012,1),to=c(2014,4),
  whichVar=c("asbhh","apkz","asbper","ajahr","aquartal","amonat",
  "xnuts2","xerwstat"))
# }