Sehr spezifische Funktion die MZ-Daten einliest um damit in weiterer Folge Tabellen im MZ-Quartalsbericht-Format zu erstellen (hausintern).

Funktion liest MZ-Daten (dg7) automatisch aus dem STAT-Filemanagement ein und fuehrt diese Daten mit dem derzeit fuer die Quartalsberichtsproduktion verwendeten SPSS-File zusammen (siehe Details).

ImportDataListQT(timeInstant, nbw = NULL, whichVar = NULL,
  weightDecimals = 2, ImportAndMerge = TRUE, curr_inFile = NULL,
  prev_inFile = NULL, mergeBy = c("asbper", "ajahr", "amonat"),
  mz_intern = mount_mz_intern(), mz_ergebnisse = mount_mz_ergebnisse())

Arguments

timeInstant
numerischer Vektor mit 2 Elementen: c(jahr, quartal). Hier gibt man den Zeitpunkt an auf den sich alle Ergebnisse im weitesten Sinn beziehen sollen, also i.d.R. das aktuellste Quartal.

nbw
numerischer Wert: Anzahl an Bootstrap-Gewichten die eingelesen werden soll (z.B. um Rechenzeit beim Aufsetzen der Tabellen zu verkuerzen).

whichVar
Character (vector) oder NULL. Falls ungleich NULL, Character Vektor mit Variable(n) aus dem dg7-Mikrozensus-File die im Output-File enthalten sein sollen. Die uebrigen Variablen werden weggelassen. Default ist NULL, dabei werden alle Variablen behalten.

weightDecimals
Numerischer Wert oder NULL. Anzahl der Nachkommastellen der Stichprobengewichte, gerundet nach SPSS RND Logik (0.5 bwz. -0.5 wird dabei immer "weg von 0" gerundet). Falls NULL, werden die Gewichte nicht gerundet.

ImportAndMerge
TRUE/FALSE ob die Funktion ImportAndMerge angewendet werden soll. Bei der Defaulteinstellung ImportAndMerge=TRUE und curr_inFile=NULL wird Daten_ab2004_QuartPub.sav zu den MZ-Daten gemerged.

curr_inFile
Pfad der Datei die eingelesen und zu den MZ-Daten gemerged werden soll (bezogen auf den aktuelleren der beiden Zeitpunkte falls prev_inFile ungleich NULL). Eingelesen werden koennen Files vom Typ .sav, .csv und .csv.gz.

prev_inFile
Falls ungleich NULL, Pfad der Datei die eingelesen und zu den MZ-Daten gemerged werden soll (bezogen auf den weniger aktuellen Zeitpunkt). Eingelesen werden koennen Files vom Typ .sav, .csv und .csv.gz.

mergeBy
Character Vektor mit Variable(n) nach denen gemerged werden soll (default=c("asbper","ajahr","amonat")).

mz_intern
Pfad zu dem mz_intern Ordner in der STAT Infrastruktur. Standardmäßig wird dieser mit mountSTAT generiert.

mz_ergebnisse
Pfad zu dem 06 Ergebnisse Ordner in der STAT Infrastruktur. Standardmäßig wird dieser mit mountSTAT generiert.

Value

Output ist eine Liste deren Elemente jeweils MZ-Daten enthalten die die selbe Grundstruktur haben wie der Output aus der Funktion ImportData.

Details

Anzugeben ist bei dieser Funktion der Referenzzeitpunkt timeInstant des MZ-Quartalsberichts. Die MZ-Daten (dg7) werden (derzeit) defaultmaessig (ImportAndMerge=TRUE) ueber mergeBy = c("asbper","ajahr","amonat") mit Daten_ab2004_QuartPub.sav gemerged. Dieses File wird quartalsweise vom Fachbereich erzeugt und immer im selben Ordner abgelegt. Sollte man mit einem anderen File mergen wollen, so kann man die entsprechenden File-Pfade in curr_inFile und prev_inFile spezifizieren aber auch mergeBy anpassen, analog zu ImportAndMerge. Des Weiteren koennen die einzulesenden MZ-Daten mit den Funktionsparametern nbw, whichVar und weightDecimals angepasst werden, siehe ImportData. Dadurch kann man z.B. vermeiden, dass Variablen die sowohl in den MZ-Daten als auch in Daten_ab2004_QuartPub.sav vorkommen doppelt eingelesen werden.

See also

Examples

# NOT RUN {
## Lesen Daten fuer den AKE-QT-Referenzzeitpunkt 2014q4 ein.
# Fuer Testzwecke (um Rechenzeit zu sparen) schraenken wir die 
# Anzahl der Bootstrapgewichte ein auf 5.
# Ausserdem wollen wir aus den Original-MZ-Daten (dg7) nur die 
# Variable rbpkin behalten.

datalist <- ImportDataListQT(timeInstant=c(2014,4), nbw=5, whichVar=c("rbpkin"))

# }