Mikrozensus-Files und zugehoerige Bootstrapgewichte einlesen.

Funktion liest vom Nutzer zur Verfuegung gestellte Datenfiles ein

IndivImportData(curr_inFile, curr_inFile_bw, prev_inFile = NULL,
  prev_inFile_bw = NULL, whichVar = NULL, mergeBy = "asbhh",
  nbw = NULL, bwNames = NULL, weightName = "gew1",
  weightDecimals = 2)

Arguments

curr_inFile
Character oder Character Vektor mit Pfad(en) der MZ-Datei(en) die eingelesen werden soll(en) (bezogen auf den aktuelleren der beiden Zeitpunkte falls prev_inFile ungleich NULL). Eingelesen werden koennen Files vom Typ .sav, .csv und .csv.gz (mit Feldtrennzeichen ; und Dezimaltrennzeichen ,). Soll ein (Mehr)Jahresdatensatz erstellt werden, muessen mehr Pfade zu den entsprechenden Quartalsfiles uebergeben werden.

curr_inFile_bw
Character oder Character Vektor mit Pfad(en) der Datei(en) mit den Bootstrapgewichten (File(s) vom Typ .csv und .csv.gz mit Feldtrennzeichen ; und Dezimaltrennzeichen ,). Soll ein (Mehr)Jahresdatensatz erstellt werden, muessen mehr Pfade zu den entsprechenden Quartalsfiles uebergeben werden.

prev_inFile
Falls ungleich NULL, Pfad(e) der MZ-Datei(en) die eingelesen werden soll(en) (bezogen auf den weniger aktuellen Zeitpunkt). Eingelesen werden koennen Files vom Typ .sav, .csv und .csv.gz (mit Feldtrennzeichen ; und Dezimaltrennzeichen ,). Soll ein(Mehr)Jahresdatensatz erstellt werden, muessen mehr Pfade zu den entsprechenden Quartalsfiles uebergeben werden.

prev_inFile_bw
Falls ungleich NULL, character oder character Vektor mit Pfad(en) der Datei(en) mit den Bootstrapgewichten (File(s) vom Typ .csv und .csv.gz mit Feldtrennzeichen ; und Dezimaltrennzeichen ,) bezogen auf den weniger aktuellen Zeitpunkt. Soll ein (Mehr)Jahresdatensatz erstellt werden, muessen 4 Pfade zu den entsprechenden Quartalsfiles uebergeben werden.

whichVar
Falls ungleich NULL, Character Vektor mit Variable(n) aus dem zur Verfuegung gestellten Datenfile die im Output-File enthalten sein sollen. Die uebrigen Variablen werden weggelassen. Default ist NULL, dabei werden alle Variablen behalten.

mergeBy
Character oder Character Vector der Variablen die zum Mergen von curr_inFile und curr_inFile_bw bzw. prev_inFile und prev_inFile_bw verwendet werden sollen. Default ist der Haushalts-Identifikator "asbhh".

nbw
Falls ungleich NULL, Anzahl an Bootstrap-Gewichten die eingelesen werden soll. Default ist NULL, dabei werden alle verfuegbaren Bootstrap-Gewichte eingelesen.

bwNames
Falls ungleich NULL, die Variablennamen der Bootstrap-Gewichte in curr_inFile_bw bzw. prev_inFile_bw. Default ist NULL, dabei ist die Variablenbezeichnung der Bootstrapgewichte "gew1_1", "gew1_2", "gew1_3", … wie beim Mikrozensus ueblich.

weightName
Character: Name des Gewichtsvektors der eingelesenen Daten, default ist weightName="gew1".

weightDecimals
Numerischer Wert oder NULL. Anzahl der Nachkommastellen der (angepassten) Stichprobengewichte, gerundet nach SPSS RND Logik (0.5 bwz. -0.5 wird dabei immer "weg von 0" gerundet). Falls NULL, werden die Gewichte nicht gerundet.

Value

Output ist eine Liste mit einem oder zwei Elementen, je nachdem ob prev_inFile=NULL oder nicht. Die Listenelemente sind Objekte der Klasse data.table. Wurden mehrere Dateipfade angegeben, so enthaelt der Output angepasste Gewichte, d.h. alle Gewichte werden durch die Anzahl der uebergebenen Dateipfade in curr_inFile dividiert. Jahresgewichte berechnen sich also als Quartalsgewichte durch 4.

Details

Das mzR-Paket benoetigt in den Daten eine Gewichtsvariable mit dem Namen "gew1" und Bootstrapgewichte mit den Namen "gew1_1", "gew1_2", "gew1_3", … (wie beim Mikrozensus ueblich). Sollten diese Variablen in den einzulesenden Daten andere Namen haben, so muessen diese in den Funktionsparametern weightName bzw. bwNames spezifiziert werden.

See also

Examples

# NOT RUN {
## Quartalsdaten
curr_inFile <- c("...pfad.../dg8.mz2014Q1.sav")
curr_inFile_bw <- c("...pfad.../mz2_2014q1_bootweights.csv.gz")
dat <- IndivImportData(curr_inFile=curr_inFile, curr_inFile_bw=curr_inFile_bw)

## Jahresdaten
curr_inFile <- c("...pfad.../dg8.mz2014Q1.sav",
                 "...pfad.../dg8.mz2014Q2.sav",
                 "...pfad.../dg8.mz2014Q3.sav",
                 "...pfad.../dg8.mz2014Q4.sav")
curr_inFile_bw <- c("...pfad.../mz2_2014q1_bootweights.csv.gz",
                    "...pfad.../mz2_2014q2_bootweights.csv.gz",
                    "...pfad.../mz2_2014q3_bootweights.csv.gz",
                    "...pfad.../mz2_2014q4_bootweights.csv.gz")
dat <- IndivImportData(curr_inFile=curr_inFile, curr_inFile_bw=curr_inFile_bw)

## Jahresdaten von 2 verschiedenen Zeitpunkten fuer Fehlerrechnung fuer Veraenderungen
curr_inFile <- c("...pfad.../dg8.mz2014Q1.sav",
                 "...pfad.../dg8.mz2014Q2.sav",
                 "...pfad.../dg8.mz2014Q3.sav",
                 "...pfad.../dg8.mz2014Q4.sav")
curr_inFile_bw <- c("...pfad.../mz2_2014q1_bootweights.csv.gz",
                    "...pfad.../mz2_2014q2_bootweights.csv.gz",
                    "...pfad.../mz2_2014q3_bootweights.csv.gz",
                    "...pfad.../mz2_2014q4_bootweights.csv.gz")
prev_inFile <- c("...pfad.../dg8.mz2013Q1.sav",
                 "...pfad.../dg8.mz2013Q2.sav",
                 "...pfad.../dg8.mz2013Q3.sav",
                 "...pfad.../dg8.mz2013Q4.sav")
prev_inFile_bw <- c("...pfad.../mz2_2013q1_bootweights.csv.gz",
                    "...pfad.../mz2_2013q2_bootweights.csv.gz",
                    "...pfad.../mz2_2013q3_bootweights.csv.gz",
                    "...pfad.../mz2_2013q4_bootweights.csv.gz")
dat <- IndivImportData(curr_inFile=curr_inFile, curr_inFile_bw=curr_inFile_bw,
                       prev_inFile=prev_inFile, prev_inFile_bw=prev_inFile_bw)
# }