Schaetz- und Fehlerrechnungsfunktionen fuer numerische Variablen.

Funktion berechnet Totalwerte oder Mittelwerte einer numerischen Variable und die zugehoerigen Fehler.

Median(x, TFstring = NULL, each = NULL, var, negativeZero = TRUE,
  thousands_separator = TRUE, digits = 2, replicates = FALSE)

Mean(x, TFstring = NULL, each = NULL, var, negativeZero = TRUE,
  thousands_separator = TRUE, digits = 2, replicates = FALSE)

Total(x, TFstring = NULL, each = NULL, var, negativeZero = TRUE,
  thousands_separator = TRUE, digits = 2, replicates = FALSE)

Arguments

x
MZ Daten - Output von Funktion ImportData.

TFstring
Character oder NULL: Logische Einschraenkung der Gruppe. Falls NULL, gilt keine Einschraenkung.

each
Character oder NULL: Name der Variable nach der getrennt berechnet werden soll. Hier koennen auch mehrere Variablen angegeben werden und zwar in einem character string, getrennt durch ein +, siehe Examples.

var
character: Name der numerischen Variable oder Formel zur Berechnung dieser z.b. 'x+y'.

negativeZero
Logical: Wenn TRUE, werden die negativen Werte der numerischen Variable auf 0 gesetzt.

thousands_separator
Logical: Wenn TRUE, werden Tausendertrennzeichen angezeigt.

digits
Numerischer Wert: Anzahl der Nachkommastellen im angezeigten Ergebnis. Default ist 2.

replicates
Fürge einen Vektor aus Schätzwerten zum Output hinzu? Die Anzahl der Schätzwerte pro Gruppe in each entspricht der Anzahl der Bootstrapreplikate (typischerweise 500). Siehe auch getReplicates.

Value

Output ist ein Objekt der Klasse mzR.

Details

Wiedergegeben wird der Schaetzer est, der Stichprobenfehler sd, der Variationskoeffizient cv und die untere/obere Grenze des 95% Konfidenzintervalls cil_2.5%/ciu_97.5%. Die Fehler werden mit Hilfe von Replikationsgewichten aus einem Bootstrapverfahren berechnet, d.h. sd entspricht der Standardabweichung der mit den Bootstrapgewichten berechneten Schaetzwerte, cil_2.5% und ciu_97.5% sind die entsprechenden 2.5% und 97.5% Quantile und cv=sd/est.

Functions

Median: Median. Robuste alternative zu Mean

Mean: Arithmetisches Mittel

See also

Examples

# Daten laden (oder Daten einlesen mit ImportData() bzw. IndivImportData())
data(mzTestData)

# Geleistete Arbeitsstunden: Absolutwerte
Total(mzTestData,TFstring="xerwstat==1&balt>=15&balt<=74",var="estund*13+dtstd*13")

#>                              est             sd    cv         cil_2.5%
#> year2           1,671,374,942.14 232,757,050.50  0.14 1,549,405,938.58
#> year1           1,829,623,967.55 252,284,180.55  0.14 1,421,018,572.91
#> Absolute change  -158,249,025.41 280,771,848.82 -1.77  -115,533,834.66
#> Relative change            -8.65          17.84 -2.06            -5.64
#>                    ciu_97.5%
#> year2           2.131948e+09
#> year1           2.001863e+09
#> Absolute change 5.126075e+08
#> Relative change 3.539000e+01

# Geleistete Arbeitsstunden: Absolutwerte in Millionen
Total(mzTestData,TFstring="xerwstat==1&balt>=15&balt<=74",var="(estund+dtstd)*13/10^6")

#>                      est     sd    cv cil_2.5% ciu_97.5%
#> year2           1,671.37 232.76  0.14 1,549.41  2,131.95
#> year1           1,829.62 252.28  0.14 1,421.02  2,001.86
#> Absolute change  -158.25 280.77 -1.77  -115.53    512.61
#> Relative change    -8.65  17.84 -2.06    -5.64     35.39

#Durchschnittlich geleistete Arbeitsstunden
Mean(mzTestData,TFstring="xerwstat==1&balt>=15&balt<=74",var="estund+dtstd")

#>                   est   sd   cv cil_2.5% ciu_97.5%
#> year2           31.08 2.50 0.08    28.46     34.70
#> year1           29.42 2.39 0.08    24.32     30.11
#> Absolute change  1.66 2.47 1.48     1.74      7.76
#> Relative change  5.66 9.51 1.68     5.92     28.86

#Durchschnittlich geleistete Arbeitsstunden nach Geschlecht
Mean(mzTestData,TFstring="xerwstat==1&balt>=15&balt<=74",var="estund+dtstd", each="bsex")

#> $bsex_1
#>                   est    sd   cv cil_2.5% ciu_97.5%
#> year2           33.36  4.32 0.13    33.38     43.27
#> year1           32.39  3.00 0.09    29.23     36.44
#> Absolute change  0.96  5.97 6.19     0.24     13.69
#> Relative change  2.98 20.65 6.94     0.71     46.94
#> 
#> $bsex_2
#>                   est    sd   cv cil_2.5% ciu_97.5%
#> year2           29.02  3.49 0.12    24.10     32.36
#> year1           25.80  7.75 0.30    10.40     29.88
#> Absolute change  3.22  4.98 1.55     2.18     13.85
#> Relative change 12.48 59.01 4.73     7.37    140.77
#> 

#Durchschnittlich geleistete Arbeitsstunden nach Bundesland und Geschlecht
Mean(mzTestData,TFstring="xerwstat==1&balt>=15&balt<=74",var="estund+dtstd",each="xnuts2+bsex")

#> $xnuts2_bsex_11001
#>                    est   sd    cv cil_2.5% ciu_97.5%
#> year2            22.00 0.00  0.00    22.00     22.00
#> year1            28.88 2.25  0.08    29.25     34.07
#> Absolute change  -6.88 2.25 -0.33   -12.07     -7.25
#> Relative change -23.81 5.01 -0.21   -35.43    -24.79
#> 
#> $xnuts2_bsex_11002
#>                   est   sd   cv cil_2.5% ciu_97.5%
#> year2           37.00 0.00 0.00    37.00     37.00
#> year1           36.25 0.34 0.01    36.25     36.92
#> Absolute change  0.75 0.34 0.45     0.08      0.75
#> Relative change  2.07 0.93 0.45     0.21      2.07
#> 
#> $xnuts2_bsex_12001
#>                     est   sd    cv cil_2.5% ciu_97.5%
#> year2             -3.00 0.00  0.00    -3.00     -3.00
#> year1             29.83 4.55  0.15    27.28     36.92
#> Absolute change  -32.83 4.55 -0.14   -39.92    -30.28
#> Relative change -110.06 1.35 -0.01  -111.01   -108.12
#> 
#> $xnuts2_bsex_12002
#>                   est     sd   cv  cil_2.5% ciu_97.5%
#> year2           31.25   1.98 0.06     29.33     33.57
#> year1           16.62  12.19 0.73     -3.00     22.51
#> Absolute change 14.63  11.70 0.80     10.76     35.33
#> Relative change 87.97 662.02 7.53 -1,177.82     91.28
#> 
#> $xnuts2_bsex_13001
#>                   est    sd   cv cil_2.5% ciu_97.5%
#> year2           42.00  0.00 0.00    42.00     42.00
#> year1           26.50 10.01 0.38    12.35     36.36
#> Absolute change 15.50 10.01 0.65     5.64     29.65
#> Relative change 58.49 98.32 1.68    15.88    249.69
#> 
#> $xnuts2_bsex_13002
#>                    est    sd    cv cil_2.5% ciu_97.5%
#> year2            14.50 15.65  1.08    -3.00     30.25
#> year1            35.50  0.00  0.00    35.50     35.50
#> Absolute change -21.00 15.65 -0.75   -38.50     -5.25
#> Relative change -59.15 44.09 -0.75  -108.45    -14.79
#> 
#> $xnuts2_bsex_21001
#>                    est    sd    cv cil_2.5% ciu_97.5%
#> year2            17.00  0.00  0.00    17.00        17
#> year1            29.50 12.50  0.42    17.00        42
#> Absolute change -12.50 12.50 -1.00   -25.00         0
#> Relative change -42.37 30.29 -0.71   -59.52         0
#> 
#> $xnuts2_bsex_21002
#>                    est     sd   cv cil_2.5% ciu_97.5%
#> year2            44.00   4.29 0.10    37.70     47.50
#> year1            20.75  16.16 0.78     6.00     35.50
#> Absolute change  23.25  15.09 0.65     8.85     41.15
#> Relative change 112.05 326.57 2.91    24.93    685.82
#> 
#> $xnuts2_bsex_22001
#>                    est    sd    cv cil_2.5% ciu_97.5%
#> year2            23.17 16.33  0.70    -2.00     34.27
#> year1            46.25  8.99  0.19    36.58     57.00
#> Absolute change -23.08 22.40 -0.97   -57.92     -2.93
#> Relative change -49.91 42.67 -0.85  -104.24     -5.94
#> 
#> $xnuts2_bsex_22002
#>                    est    sd    cv cil_2.5% ciu_97.5%
#> year2            23.67  9.43  0.40    12.50     35.75
#> year1            37.00  0.00  0.00    37.00     37.00
#> Absolute change -13.33  9.43 -0.71   -24.50     -1.25
#> Relative change -36.04 25.48 -0.71   -66.22     -3.38
#> 
#> $xnuts2_bsex_31001
#>                   est    sd   cv cil_2.5% ciu_97.5%
#> year2           40.33  4.47 0.11    37.50     47.00
#> year1           36.25  0.34 0.01    35.58     36.25
#> Absolute change  4.08  4.18 1.02     1.93     10.75
#> Relative change 11.26 11.50 1.02     5.39     29.66
#> 
#> $xnuts2_bsex_31002
#>                   est       sd        cv cil_2.5% ciu_97.5%
#> year2           37.25     7.76      0.21    32.30     51.05
#> year1           37.67    19.84      0.53     1.00     37.60
#> Absolute change -0.42    26.75    -64.19    -4.77     50.05
#> Relative change -1.11 2,557.79 -2,312.24   -12.87  5,003.99
#> 
#> $xnuts2_bsex_32001
#>                    est    sd    cv cil_2.5% ciu_97.5%
#> year2            29.88  9.73  0.33    20.54     43.17
#> year1            41.83  3.58  0.09    37.00     45.13
#> Absolute change -11.96 11.60 -0.97   -20.01      6.02
#> Relative change -28.59 28.51 -1.00   -49.56     16.28
#> 
#> $xnuts2_bsex_32002
#>                    est    sd   cv cil_2.5% ciu_97.5%
#> year2            40.50  1.92 0.05    40.50     44.00
#> year1            14.00  0.00 0.00    14.00     14.00
#> Absolute change  26.50  1.92 0.07    26.50     30.00
#> Relative change 189.29 13.69 0.07   189.29    214.29
#> 
#> $xnuts2_bsex_33001
#>                   est    sd   cv cil_2.5% ciu_97.5%
#> year2           57.88  4.71 0.08    54.18     64.50
#> year1           33.90  4.31 0.13    30.17     40.33
#> Absolute change 23.98  7.45 0.31    15.20     34.13
#> Relative change 70.72 30.36 0.43    38.91    114.06
#> 
#> $xnuts2_bsex_33002
#>                    est    sd    cv cil_2.5% ciu_97.5%
#> year2             -3.0  0.00  0.00    -3.00     -3.00
#> year1             12.5  3.13  0.25     5.25     12.20
#> Absolute change  -15.5  3.13 -0.20   -15.20     -8.25
#> Relative change -124.0 15.69 -0.13  -157.14   -124.76
#> 
#> $xnuts2_bsex_34001
#>                   est   sd    cv cil_2.5% ciu_97.5%
#> year2           38.00 0.89  0.02    38.20     40.00
#> year1           38.67 1.18  0.03    37.00     39.42
#> Absolute change -0.67 1.16 -1.75     0.55      3.00
#> Relative change -1.72 3.18 -1.85     1.41      8.11
#> 
#> $xnuts2_bsex_34002
#>                    est     sd   cv cil_2.5% ciu_97.5%
#> year2            16.00   9.84 0.61     5.00     25.90
#> year1             7.00   8.94 1.28    -2.00     17.00
#> Absolute change   9.00   8.82 0.98   -11.00      9.80
#> Relative change 128.57 123.32 0.96  -247.07     52.35
#>