isolationWindow-methods {mzR}	R Documentation

Returns the ion selection isolation window

Description

The methods return matrices of lower (column low) and upper (column high) isolation window offsets. Matrices are returned as a list of length equal to the number of input files (provided as file names of raw mass spectrometry data objects, see below). By default (i.e when unique. = TRUE ), only unique offsets are returned, as they are expected to identical for all spectra per acquisition. If this is not the case, a message is displayed.

Methods

signature(object = "character", unique. = "logical", simplify = "logical")

Returns the isolation window for the file object. By default, only unique isolation windows are returned per file (unique = TRUE); if set to FALSE, a matrix with as many rows as there are MS2 spectra. If only one file passed an input and simplify is set to TRUE (default), the resulting list of length 1 is simplified to a matrix.

signature(object = "mzRpwiz", unique. = "logical", simplify = "logical")

As above for mzRpwiz objects.

signature(object = "mzRramp", unique. = "logical", simplify = "logical")

As above for mzRramp object.

Author(s)

Laurent Gatto <lg390@cam.ac.uk> based on the functionality from the msPurity:::get_isolation_offsets function.

Examples

library("msdata")
f <- msdata::proteomics(full.names = TRUE,
                        pattern = "TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01.mzML.gz")
isolationWindow(f)

rw <- openMSfile(f)
isolationWindow(rw)
str(isolationWindow(rw, unique = FALSE))

[Package mzR version 2.18.0 Index]