metadata {mzR}	R Documentation

Access the metadata from an mzR object.

Description

Accessors to the analytical setup metadata of a run. runInfo will show a summary of the experiment as a named list, including scanCount, lowMZ, highMZ, dStartTime, dEndTime and startTimeStamp. Note that startTimeStamp can only be extracted from mzML files using the pwiz backend or from CDF files. A NA is reported if its value is not available. The instrumentInfo method returns a named list including instrument manufacturer, model, ionisation technique, analyzer and detector. mzRpwiz will give more additional information including information on sample, software using and original source file. These individual pieces of information can also be directly accessed by the specific methods. mzidInfo is used for the mzR object created from a mzid file. It returns basic information on this mzid file including file provider, creation date, software, database, enzymes and spectra data format. The mzidInfo will return the scoring results in identification. It will return different results for different searching software used.

Usage

runInfo(object)
chromatogramsInfo(object)
analyzer(object)
detector(object)
instrumentInfo(object)
ionisation(object)
softwareInfo(object)
sampleInfo(object)
sourceInfo(object)
model(object)
mzidInfo(object)
modifications(object, ...)
psms(object, ...)
substitutions(object)
database(object, ...)
enzymes(object)
tolerance(object)
score(x, ...)
para(object)
specParams(object)

Arguments

object
An instantiated mzR object.

x
An instantiated mzR object.

...
Additional arguments, currently ignored.

Author(s)

Steffen Neumann, Laurent Gatto and Qiang Kou

See Also

See for example peaks to access the data for the spectra in a "mzR" class.

Examples

 library(msdata)

 file <- system.file("microtofq/MM8.mzML", package = "msdata")
 mz <- openMSfile(file)
 fileName(mz)
 instrumentInfo(mz)
 runInfo(mz)
 close(mz)

 file <- system.file("cdf/ko15.CDF", package = "msdata")
 mz <- openMSfile(file, backend = "netCDF")
 fileName(mz)
 instrumentInfo(mz)
 runInfo(mz)
 close(mz)
 
 file <- system.file("mzid", "Tandem.mzid.gz", package="msdata")
 mzid <- openIDfile(file)
 softwareInfo(mzid)
 enzymes(mzid)

[Package mzR version 2.18.0 Index]