mzR: Microcensus tools

R package to compute estimates and their errors for Austrian Microcensus (labour force survey) data.

Test data

mzTestData

Import of data from STAT internal sources

Import data from the file system provided by STAT. These functions will not be operational for external users.

ImportData, ImportDataListQT, ImportAndMerge

Individual data import

Import spss and csv files based on custom paths.

IndivImportData

Compute estimates and uncertainty

Create objects of the class mzR which contain estimates, confidence intervals and more based on calibrated bootstrapping.

GroupRate, GroupSize, Mean, Total, Median

Methods for class mzR

print.mzR, as.data.frame.mzR, plot.mzR, AddVariable, GetLabels

Calculate tables

Calculate tables for excel export

MakeTable, MakeQT, MakeAKETimeInstantsTable

Excel export

export, FillExcelTemplate