mzR-class {mzR}	R Documentation

Class mzR and sub-classes

Description

The class mzR is the main class for the common mass spectrometry formats. It is a virtual class and thus not supposed to be instanciated directly.

The sub-classes implement specific APIs to access the underlying data and metadata in the files. Currently, mzRramp and mzRpwiz are available implementations. mzRramp uses the ISB 'RAMP' random access C/C++ API, and mzRpwiz uses Proteowizard to access the relevant information in mzData, mzXML and mzML files. mzRident is used as an interface to mzIdentML files.

IMPORTANT: New developers that need to access and manipulate raw mass spectrometry data are advised against using this infrastucture directly. They are invited to use the corresponding MSnExp (with on disk mode) from the MSnbase package instead. The latter supports reading multiple files at once and offers access to the spectra data (m/z and intensity) as well as all the spectra metadata using a coherent interface. The MSnbase infrastructure itself used the low level classes in mzR, thus offering fast and efficient access.

Objects from the Class

mzR is a virtual class, so instances cannot be created.

Objects can be created by calls of the form new("mzRramp", ...), but more often they will be created with openMSfile.

After creating an mzR object, one can write it into a new file. mzXML, mzML, mgf formats are supported.

Slots

fileName:

Object of class character storing the original filename used when the instance was created.

backend:

One of the implemented backens or NULL.

.__classVersion__:

Object of class "Versioned", from Biobase.

Extends

Class "Versioned", directly.

Methods

For methods to access raw data (spectra and chromatograms), see peaks.

Methods currently implemented for mzR

fileName

signature(object = "mzR"): ...

Methods currently implemented for mzRramp

analyzer

signature(object = "mzRramp"): ...

close

signature(con = "mzRramp"): ...

detector

signature(object = "mzRramp"): ...

fileName

signature(object = "mzRramp"): ...

initializeRamp

signature(object = "mzRramp"): ...

instrumentInfo

signature(object = "mzRramp"): ...

ionisation

signature(object = "mzRramp"): ...

isInitialized

signature(object = "mzRramp"): ...

length

signature(x = "mzRramp"): ...

manufacturer

signature(object = "mzRramp"): ...

model

signature(object = "mzRramp"): ...

runInfo

signature(object = "mzRramp"): ...

Methods currently implemented for mzRpwiz

analyzer

signature(object = "mzRpwiz"): ...

detector

signature(object = "mzRpwiz"): ...

instrumentInfo

signature(object = "mzRpwiz"): ...

ionisation

signature(object = "mzRpwiz"): ...

length

signature(x = "mzRpwiz"): ...

manufacturer

signature(object = "mzRpwiz"): ...

model

signature(object = "mzRpwiz"): ...

runInfo

signature(object = "mzRpwiz"): ...

chromatogramsInfo

signature(object = "mzRpwiz"): ...

Methods currently implemented for mzRident

mzidInfo

signature(object = "mzRident"): ...

psms

signature(object = "mzRident"): ...

modifications

signature(object = "mzRident"): ...

substitutions

signature(object = "mzRident"): ...

database

signature(x = "mzRident"): ...

enzymes

signature(object = "mzRident"): ...

sourceInfo

signature(object = "mzRident"): ...

tolerance

signature(object = "mzRident"): ...

score

signature(object = "mzRident"): ...

para

signature(object = "mzRident"): ...

specParams

signature(object = "mzRident"): ...

Author(s)

Steffen Neumann, Laurent Gatto, Qiang Kou

References

RAMP: http://tools.proteomecenter.org/wiki/index.php?title=Software:RAMP Proteowizard: http://proteowizard.sourceforge.net/

Examples

 library(msdata)
 filepath <- system.file("microtofq", package = "msdata")
 file <- list.files(filepath, pattern="MM14.mzML",
		     full.names=TRUE, recursive = TRUE)
 mzml <- openMSfile(file)
 close(mzml)

 ## using the pwiz backend
 mzml <- openMSfile(file, backend = "pwiz")

[Package mzR version 2.20.0 Index]