Testdata resembling a subsample of the Austrian Microcensus
Source: R/data.R

A list containing two datasets corresponding to the time points year 1 and year 2. Each dataset consists of a selection of variables from a subset of the Austrian Microcensus including weights and five bootstrap weights.

mzTestData

Format

A list containing two datasets of class data.table with 90 rows and 13 variables each. The variables are as follows:

asbper: personal ID

bsex: gender (1 = male, 2 = female)

xnuts2: NUTS-2 region, i.e. Austrian federal states (11 = Burgenland, 12 = Lower Austria, 13 = Vienna, 21 = Carinthia, 22 = Styria, 31 = Upper Austria, 32 = Salzburg, 33 = Tyrol, 34 = Vorarlberg)

xerwstat: activity status during the reference week (1 = in employment, 2 = unemployed, 3 = economically inactive, 4 = compulsory military or community service, 9 = not applicable (child less than 15 years))

balt: age

estund: number of hours actually worked during the reference week in the second job

dtstd: number of hours actually worked during the reference week in the main job

gew1: weights

gew1_1, …, gew1_5: bootstrap weights