openMSfile {mzR}	R Documentation

Create and check mzR objects from netCDF, mzXML, mzData or mzML files.

Description

The openMSfile constructor will create a new format-specifc mzR object, open 'filename' file and all header information is loaded as a Rcpp module and made accessible through the ramp or pwiz slot of the resulting object.

The openIDfile constructor will create a new format-specifc mzR object, open 'filename' file and all information is loaded as a Rcpp module. The mzid format is supported throught pwiz backend. Only mzIdentML 1.1 is supported.

Usage

 openMSfile(filename, backend = NULL, verbose = FALSE)

 initializeRamp(object)

 isInitialized(object)

 fileName(object, ...)
 
 openIDfile(filename, verbose = FALSE)

Arguments

filename
Path name of the netCDF, mzData, mzXML or mzML file to read/write.

backend
A character(1) specifiying which backend API to use. Currently 'Ramp', 'netCDF' and 'pwiz' are supported. If backend = NULL (the default), the function tries to determine the backend to be used based on either the file extension of the file content.

object
An instantiated mzR object.

verbose
Enable verbose output.

...
Additional arguments, currently ignored.

Author(s)

Steffen Neumann, Laurent Gatto, Qiang Kou

Examples

 library(msdata)
 filepath <- system.file("microtofq", package = "msdata")
 file <- list.files(filepath, pattern="MM14.mzML",
                     full.names=TRUE, recursive = TRUE)
 mz <- openMSfile(file)
 fileName(mz)
 runInfo(mz)
 close(mz)

 ## Not run: 
    ## to use another backend
    mz <- openMSfile(file, backend = "pwiz")
    mz
  
## End(Not run)
  
  file <- system.file("mzid", "Tandem.mzid.gz", package="msdata")
  mzid <- openIDfile(file)
  softwareInfo(mzid)
  enzymes(mzid)

[Package mzR version 2.18.0 Index]